refactor: migrateWidgets
